requirements.txt
Django==5.2.5
psycopg2-binary==2.9.10
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-environ==0.11.2
Pillow==10.2.0
django-filter==23.5
drf-yasg==1.21.7
coverage==7.4.1
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
mozilla-django-oidc==4.0.1
django-allauth==0.60.1
pika==1.3.2
africastalking==1.2.8
django-model-utils==4.3.1
